+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Essential Skills Coach Dashboard [#3425](https://github.com/rokwire/illinois-app/issues/3425).
 - Primary care provider and academic advisors [#3925](https://github.com/rokwire/illinois-app/issues/3925).
+### Changed
+- Group event visibility wording [#3906](https://github.com/rokwire/illinois-app/issues/3906).
 
 ## [5.1.18] - 2024-01-31
 ### Fixed